Add files via upload
This repository contains code for 4 modules of a Dynamic Traffic Management System with Override System for Emergency Vehicles. 
A model was trained using YOLO for detecting ambulances. The weights for the model has not been included in this. 
All the videos and images along with Module wise code divisions are provided.